Update types of scores

+++ src/state/include/state/interfaces/i_command_taker.h
@@ -71,7 +71,7 @@ class STATE_EXPORT ICommandTaker : public IUpdatable {
      *
      * @return scores
      */
-    virtual const std::array<size_t, 2> getScores(bool game_over) const = 0;
+    virtual const std::array<int64_t, 2> getScores(bool game_over) const = 0;
 
     /**
      * Check if the game is over
